+++ R/ped_spatial.R
@@ -84,9 +84,9 @@
 #' org_tables <- org_fams(ped_colony, sampledata, output = "both")
 #' # Prepare data for plotting.
 #' pt <- plot_table(plot_fams = 1,
-#'   org_tables$fams,
-#'   org_tables$ped,
-#'   sampledata,
+#'   all_fams = org_tables$fams,
+#'   ped = org_tables$ped,
+#'   sampledata = sampledata,
 #'   deadSample = c("Tissue")
 #' )
 #'