+++ tests/testthat/test-ped_spatial.R
@@ -9,9 +9,9 @@ ped_colony <- get_colony(path, sampledata, rm_obsolete_parents = TRUE, out = "Fa
 
 org_tables <- org_fams(ped_colony, sampledata, output = "both")
 pt <- plot_table(plot_fams = 1,
-                 org_tables$fams,
-                 org_tables$ped,
-                 sampledata,
+                 all_fams = org_tables$fams,
+                 ped = org_tables$ped,
+                 sampledata = sampledata,
                  deadSample = c("Tissue", "Decomposing Tissue", "Blood")
 )