+++ tests/testthat/test-plot_table.R
@@ -7,9 +7,9 @@ path <- paste0(system.file("extdata", package = "wpeR"), "/wpeR_samplePed")
 ped_colony <- get_colony(path, sampledata, rm_obsolete_parents = TRUE, out = "FamAgg")
 org_tables <- org_fams(ped_colony, sampledata, output = "both")
 pt <- plot_table(plot_fams = "all",
-           org_tables$fams,
-           org_tables$ped,
-           sampledata,
+           all_fams = org_tables$fams,
+           ped = org_tables$ped,
+           sampledata = sampledata,
            deadSample = c("Tissue", "Decomposing Tissue", "Blood"))
 
 
@@ -40,5 +40,22 @@ test_that("all reproductive animals are flagged", {
   expect_true(all(unique(pt$AnimalRef[pt$rep==TRUE]) %in% unlist(org_tables$fams[c(2,3)], use.names = FALSE)))
 })
 
+test_that("plot_indivs filters families correctly", {
+  # Pick an individual
+  indiv <- org_tables$fams$mother[1]
 
+  pt_indiv <- plot_table(all_fams = org_tables$fams,
+                         plot_indivs = indiv,
+                         ped = org_tables$ped,
+                         sampledata = sampledata,
+                         )
 
+  # Find all families she is associated with (as mother, father, or offspring)
+  all_associated_fams <- unique(c(
+      org_tables$fams$FamID[org_tables$fams$father == indiv | org_tables$fams$mother == indiv],
+      org_tables$ped$FamID[org_tables$ped$id == indiv]
+  ))
+
+  # Check if all families in pt_indiv are associated with this individual
+  expect_true(all(unique(pt_indiv$FamID) %in% all_associated_fams))
+})