+++ tests/testthat/test-pps_subfunctions.R
@@ -7,9 +7,9 @@ path <- paste0(system.file("extdata", package = "wpeR"), "/wpeR_samplePed")
 ped_colony <- get_colony(path, sampledata, rm_obsolete_parents = TRUE, out = "FamAgg")
 org_tables <- org_fams(ped_colony, sampledata, output = "both")
 pt <- plot_table(plot_fams = "all",
-                 org_tables$fams,
-                 org_tables$ped,
-                 sampledata,
+                 all_fams = org_tables$fams,
+                 ped = org_tables$ped,
+                 sampledata = sampledata,
                  deadSample = c("Tissue", "Decomposing Tissue", "Blood"))
 pt$AnimalRef_Fam <- paste0(pt$AnimalRef, "_", pt$FamID)