+++ NEWS.md
@@ -1,3 +1,9 @@
+# wper 0.2.0 —> need to bump
+* documentation edits for clarifications
+* plot_table() added functionality to filter families based on indiviudal ids
+(parameter plot_indivs )
+* ped_satplot() added new visual representation for polygamous individuals
+
 # wpeR 0.1.1
 
 * fixed bug in spatial function that creates polygons